scripts/requirements.txt
Flask
gspread
pandas
google-auth
google-auth-oauthlib
google-api-python-client
gunicorn
google-cloud-storage
xlrd
openpyxl
PyMuPDF # Add this line for PDF processing